Update patch set 4
Patch Set 4: Code-Review+1

> AAI-57 says that the auth library will be part of the gizmo repo,
 > is this intended for aai-common?  Should the story be changed?

The auth library is going to be used by Gizmo, but it is intended as a common library which can also be used by other services.

Based on the the thread on the discussion group, the consensus seemed to be that we wanted to put this type of library into the aai-common repository.

So yes, maybe the wording in AAI-57 needs to be changed.

Patch-set: 4
Reviewer: Gerrit User 453 <453@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1