Update patch set 8
Patch Set 8:

Network Service is a concept from ETSI IFA. It is composed of VNFs, VL and VNFFG(Future) resources. In the VoLTE case, there is E2E service, which is composed of EPC network service and IMS network service in NFV domain, and underlay/overlay VPN in the SDN domain. Network service's attributes are different from E2E service or service from the openECOMP.
If support the two layer services in AAI, it will more clear for the code.

Patch-set: 8