Update patch set 8
Patch Set 8:

If these objects are not modelled in SDC, aai prefers the use of service-instance; service-type is already an attribute on the service-instance object. we can add a status attribute to service-instance unless orchestration-status is sufficient. we can add createdAt and updateAt attrs if needed to service-instance.  If the object is modelled in SDC, it seems like we should add the model attributes (at minimum, model-invariant-id and model-version-id)

Patch-set: 8