Update patch set 2
Patch Set 2:

Jimmy, this was an accidental mixup. The configurationId must be the first element but the management option which isn't a key is the first element. We noticed that if the key isn't the first element, there are issues with it.

Patch-set: 2
Reviewer: Gerrit User 75 <75@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0