Update patch set 1
Patch Set 1:

> > > (1 comment)
 > > >
 > > > Please note that we have not included the TSC section.
 > > > All approval and removal is done in the wiki.onap.org and it is
 > > > approved by Gildas:
 > > > https://wiki.onap.org/pages/viewpage.action?pageId=11929094
 > >
 > > There should be some sort of non-mutable public paper trail for
 > > this and _not_ a wiki page and the approvals section, while
 > > optional, _should_ be included with that information!
 > 
 > @Andy, as the TSC has officially created the role of
 > "Infrastructure Coordinator" and nominated myself for that
 > position, for traceability purpose I propose a central wiki page
 > with write permission to myself only to record the request and my
 > approval. Thoughts?

Our paper trail _must_ be non-mutable. A wiki page does not qualify. You should be reporting to the TSC your authorizations and those should be getting reported either on the mailing list or read into the TSC meeting minutes each week. In any case it has to go to something that cannot be changed after the fact.

Patch-set: 1