Update patch set 1
Patch Set 1:

> (1 comment)

@Jimmy, We need some traceability on addition and removal of committers. The process is defined at https://wiki.onap.org/pages/viewpage.action?pageId=11929094
It may very well be that a Committer simply do not wish to contribute (for whatever reason). In that case, make a copy of the template in the wiki, upload any email you have from Steve and I let officially inform the TSC (and I will include a link of that email TSC evidence).

Patch-set: 1