Update patch set 1
Patch Set 1: Code-Review-1

I would suggest to put it on hold since we need to figure out if all HPA information (even for those applicable in Beijing Release) can be populated as child object of any flavor.
My suggestion is to place all HPA information as child object of a cloud region, and to associate those HPA pertaining to flavors to flavor objects.
With this association, OOF can still work the way as it is now.

Patch-set: 1
Label: Code-Review=-1