Update patch set 4
Patch Set 4: Code-Review+1

(6 comments)

As my repeated comment states, this really is a personal preference here, but I prefer to see subkeys indented in YAML under their parent key. I know that YAML does not specifically require this, but it makes it more apparent to people reading that something is a subkey.

That being said, I believe everything else is in order so I'll give a +1 as there is nothing actually wrong with anything I see now.

Patch-set: 4
Label: Code-Review=+1