Update patch set 7
Patch Set 7:

Dileep, can you try to capture what you are trying to do to the schema in a unit test. For example, we have a junit that creates a basic pserver and a cloud region and creates a relationship between them. In order to properly verify what you are trying to do makes sense, it will be good to have this test case captured.

JUNIT
https://git.onap.org/aai/aai-common/tree/aai-core/src/test/java/org/onap/aai/rest/PserverTest.java

Resources
https://git.onap.org/aai/aai-common/tree/aai-core/src/test/resources/payloads/templates/pserver.json
https://git.onap.org/aai/aai-common/tree/aai-core/src/test/resources/payloads/templates/relationship.json

Patch-set: 7
Label: Code-Review=0