Update patch set 2
Patch Set 2:

On the meaning of networkRef, I will seek clarification from developers.

On indexing of transparent, it was a guess that it would help with query filters, but it can be omitted if necessary.

On inv-status vs operational-status, we don't know the full impact of reusing/renaming an existing status field, since that affects other components, so it seemed prudent to add a new status to retain backwards-compatibility and limit the scope of breaking changes.

As to the naming of "orchestration-status" vs "operational-status", please note that "operational-status" fields have also been added to several other elements (commits not yet pushed) so that will have impact across the code for several elements.

Patch-set: 2