Update patch set 2
Patch Set 2:

Not sure that sp-partner was intended to store the credentials. There is another commit coming that defines the external ONAP with relationship to esr-system-info. The commit that is coming soon agrees with you.

Will ask the developers about the "callsource" property.

sp-partner did not seem like a "network" object to me, so best guess is "business" fits, unless a better suggestion is made.

Patch-set: 2