Update patch set 3
Patch Set 3:

(1 comment)

Keong, the problem isn't with the private edge unit tests. Its that the java-attribute has to be camel case. The unit tests use the schema and tries to create the schema and it failed during that phase and the first junit that failed with it was the private edge. You have cir, eir, ebs, cbs as uppercase for the java attribute which you cannot do.

Patch-set: 3
Reviewer: Gerrit User 75 <75@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0