Update patch set 4
Patch Set 4:

The "complex" object appears to be tied to the "cloud-infrastructure" and "data centre" but not sure that part is relevant to this usecase.
Will check with developers if it's OK to substitute "complex" instead.

Patch-set: 4
Reviewer: Gerrit User 4341 <4341@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0