Update patch set 1
Patch Set 1: Code-Review-1

This is testing some very basic functionality of the RestHandlerService class, and will help increase code coverage.

However, the intention of the code in the RestHandlerService class is to maintain a singleton instance of itself, so a more useful test would be to assert that calling the getInstance() method twice will return the same instance both times.

The intention of writing unit tests should not be purely for increasing test coverage stats (the stats serve as an indication of the current state of testing, not as an end goal). Instead, unit tests should serve to test business functionality of the code. For example as the case is here, the fact that the singleton pattern is doing its job correctly.

I suggest refactoring the test to verify singleton behaviour, and not the number of threads in the thread pool.

Patch-set: 1
Reviewer: Gerrit User 703 <703@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1