Update patch set 2
Patch Set 2:

Yes, these rules were both existing in v13 Beijing release before the files were re-factored.
It is simply restoring what was lost in the re-factor. The positions of the original EdgeRules can be seen in the screenshots of the WinMerge as attached to JIRA.
I suspect that if this is cherry-picked into Dublin, it will also require a v15 version too.

Patch-set: 2