Update patch set 1
Patch Set 1:

Thanks Keong, I missed that fact.

So correct me if I'm wrong, endpoints are generated from .xsd and indeed are needed inside JAR. On the other hand .html and .yaml are only for docs purpose, so they shouldn't be packaged in JAR at all.

Hence we should move them to /target or other excluded folder. 

But in the end my goal is to not have dandling files after every maven build.

Patch-set: 1