Update patch set 1
Patch Set 1:

For the generated HTML, I think this is only for documentation purpose. There still needs to be a way to distribute the files, otherwise they are pretty useless to other developers. Packaging inside the JAR is probably a least-worst option for now.

Not sure whether the YAML files are used as input to some other process. I suspect they need to go with the XSD files and POJO codes that are generated so that some clients can interface with AAI.

I agree the dangling files after a build are annoying (there are actually more files generated from the test run as well), since unclean working directories block some git operations.

I think they should generate into the target directory.

Patch-set: 1