Update patch set 1
Patch Set 1:

> [...] this is only for documentation purpose. There still needs to be a way to distribute the files. [...] Packaging inside the JAR is probably a least-worst option for now.

Like every docs they should be distributed in some way but they are not functional part of JAR. Regular folder like aai-common/docs is perfect place for such things.

I investigated a little bit and this is not that easy. aai-schema is not included as dependency per se (at least in aai-resources scope).

aai-resources uses maven-dependency-plugin to extract oxm/* and dbedgerules/* files from actual dependency. Other files seems to be ignored. Hence it doesn't matter if these files are inside JAR because they are effectively ignored.

This is a twisted solution, it could be simplified but this involves a way more changes than simple .gitignore which solves original problem.

Patch-set: 1