Update patch set 1
Patch Set 1:

The AAI Schema Service will slowly untwist the implementation and dependencies, but that is not completed yet.
The biggest problem with the Swagger HTML as part of the docs is that it is tightly coupled to the version of the code, so it will change frequently during development.
The benefit of packaging with the JAR file is that it should match the code version during the time that developers of clients of AAI will need to refer to new feature documentation.

Patch-set: 1