Update patch set 4
Patch Set 4:

(3 comments)

The versions change in this repo will only going to impact the aai-common repo junits. There is a story to retire the old v8 v9 versions. I was planning to set the ground work here. Once we communicated that we are planning to retire the changes, I will update the application properties in oom and heat repos to retire those versions.

Patch-set: 4