Update patch set 2
Patch Set 2:

In fact the changes in patch 2 may not be correct for the latest version of Eclipse, as this causes the line indentation to be 8 spaces (2 times 4). The Google style guide states "Indent continuation lines at least +4 spaces".
It probably makes sense for the ONAP standards to explicitly state how many spaces to indent.

Patch-set: 2