Fix wrapped line formatting
Update the setting "Default indentation for wrapped lines" to be 2 units
of indentation (as per the standard Google style guide formatter).
This causes a continued line to be indented differently to the following
code block and distinguishes a wrapped line from a non-wrapped line.

Also update the array initializer indentation to match, so that array
values are horizontally aligned with other initializers such as List
contructors and also with standard expressions.

Both settings are now identical with the Google-supplied formatter.

Change-Id: I05dd2b3ff522c75bb6aaf2c03017874d42e215c7
Issue-ID: AAI-2198
Signed-off-by: mark.j.leonard <mark.j.leonard@gmail.com>