Draft schema abstraction lib for review
Introduction of a schema abstraction library which will implement an
interface by which clients can query the schema without needing to
understand the complexity of loading and parsing OXM.  The library also
abstracts away the concept of OXM such that OXM could be replaced by
another format for defining the schema without the calling client being
aware.

Part of this submission includes the beginnings of an alternative schema
format (json instead of OXM).  That implementation is found in the
schemaif.json package.  It is very much a work in progress.

Issue-ID: AAI-2394
Change-Id: I39101ca03c1b3231df52e472b4e1bc4928d12f38
Signed-off-by: sblimkie <steven.blimkie@amdocs.com>