Update patch set 2
Patch Set 2:

Hi Jimmy,

We will be thankful if you can approve this review.

Also as discussed in meeting, we have raised this review request, but as aai-common is used in various microservices so those would also need validation before this is merged.

Therefore I recommend this to be reviewed, but NOT merged until all microservices are upgraded.

Thanks and regards,
Sneha Dantkale

Patch-set: 2